changed who the user will interact with the library
