+++ dynamicdevicedisplay/build.gradle
@@ -4,7 +4,6 @@ android {
     compileSdkVersion 28
 
 
-
     defaultConfig {
         minSdkVersion 26
         targetSdkVersion 28
