+++ dynamicdevicedisplay/src/main/AndroidManifest.xml
@@ -1,2 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
-<manifest package="me.aidengaripoli.dynamicdevicedisplay" />
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    package="me.aidengaripoli.dynamicdevicedisplay">
+
+    <application>
+        <activity android:name=".DeviceDiscoveryActivity"></activity>
+        <activity android:name=".DeviceActivity" />
+    </application>
+
+</manifest>
